remove usage of WTNZ name
