+++ README.md
@@ -3,4 +3,4 @@
 [![CodeFactor](https://www.codefactor.io/repository/github/amphaal/feedtnz/badge)](https://www.codefactor.io/repository/github/amphaal/feedtnz)
 [![Codacy Badge](https://app.codacy.com/project/badge/Grade/b569eb64104e42589fd8825098562243)](https://www.codacy.com/gh/Amphaal/feedtnz/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=Amphaal/feedtnz&amp;utm_campaign=Badge_Grade)
 
-Small companion app for desktop to interact with WTNZ
+Small companion app for desktop to feed or stream ITunes / Music library informations
