+++ templates/_version.h
@@ -6,3 +6,4 @@
 #define APP_FULL_DENOM "${PROJECT_NAME} ${CMAKE_PROJECT_VERSION}"
 #define APP_REMOTE_MANIFEST_URL "${APP_REMOTE_MANIFEST_URL}"
 #define APP_PATCHNOTE_URL "${APP_PATCHNOTE_URL}"
+#define DEST_PLATFORM_PRODUCT_NAME "${DEST_PLATFORM_PRODUCT_NAME}"
